Add documentation, and fix up Match
