+++ docs/src/SUMMARY.md
@@ -14,10 +14,12 @@
     - [Compress](./filters/compress.md)
     - [Concatenate Bytes](./filters/concatenate_bytes.md)
     - [Debug](./filters/debug.md)
+    - [Drop](./filters/drop.md)
     - [Firewall](./filters/firewall.md)
     - [Load Balancer](./filters/load_balancer.md)
     - [Local Rate Limit](./filters/local_rate_limit.md)
-    - [Matches](./filters/matches.md)
+    - [Match](./filters/match.md)
+    - [Pass](./filters/pass.md)
     - [Token Router](./filters/token_router.md)
     - [Writing Custom Filters](./filters/writing_custom_filters.md)
 - [Integrations](./integrations.md)
