+++ docs/src/filters/match.md
@@ -6,10 +6,11 @@ in Rust or `switch` statements in other languages.
 
 #### Filter name
 ```text
-quilkin.extensions.filters.matches.v1alpha1.Matches
+quilkin.extensions.filters.match.v1alpha1.Match
 ```
 
 ### Configuration Examples
+<!-- ANCHOR: example -->
 ```rust
 # let yaml = "
 version: v1alpha1
@@ -24,26 +25,25 @@ static:
           metadataKey: myapp.com/token
           size: 3
           remove: false
-    - name: quilkin.extensions.filters.matches.v1alpha1.Matches
+    - name: quilkin.extensions.filters.match.v1alpha1.Match
       config:
           on_read:
             metadataKey: myapp.com/token
             branches:
                 - value: abc
-                  filter: quilkin.extensions.filters.concatenate_bytes.v1alpha1.ConcatenateBytes
-                  config:
-                    on_read: APPEND
-                    bytes: eHl6 # "xyz"
+                  filter: quilkin.extensions.filters.pass.v1alpha1.Pass
+            fallthrough: quilkin.extensions.filters.drop.v1alpha1.Drop
 # ";
 # let config = quilkin::config::Config::from_reader(yaml.as_bytes()).unwrap();
 # assert_eq!(config.source.get_static_filters().unwrap().len(), 1);
 # quilkin::Builder::from(std::sync::Arc::new(config)).validate().unwrap();
 ```
+<!--  ANCHOR_END: example -->
 
-### Configuration Options ([Rust Doc](../../api/quilkin/filters/matches/struct.Config.html))
+### Configuration Options ([Rust Doc](../../api/quilkin/filters/match/struct.Config.html))
 
 ```yaml
-{{#include ../../../target/quilkin.extensions.filters.matches.v1alpha1.yaml}}
+{{#include ../../../target/quilkin.extensions.filters.match.v1alpha1.yaml}}
 ```
 
-View the [Matches](../../api/quilkin/filters/matches/struct.Config.html) filter documentation for more details.
+View the [Matches](../../api/quilkin/filters/match/struct.Config.html) filter documentation for more details.
