+++ docs/src/filters/pass.md
@@ -1,0 +1,20 @@
+# Pass
+
+The `Pass` filter that always passes any packet through it. This is useful in
+combination with conditional flow filters like [`Matches`](./match.md)
+
+#### Filter name
+```text
+quilkin.extensions.filters.pass.v1alpha1.Pass
+```
+
+### Configuration Examples
+{{#include match.md:example}}
+
+### Configuration
+
+No defined configuration options.
+
+### Metrics
+
+This filter currently exports no metrics.
