+++ proto/quilkin/extensions/filters/drop/v1alpha1/drop.proto
@@ -18,8 +18,6 @@ syntax = "proto3";
 
 package quilkin.extensions.filters.drop.v1alpha1;
 
-import "google/protobuf/wrappers.proto";
-
 message Drop {
 }
