+++ src/filters/drop.rs
@@ -64,6 +64,10 @@ impl FilterFactory for DropFactory {
         NAME
     }
 
+    fn config_schema(&self) -> schemars::schema::RootSchema {
+        schemars::schema_for!(Config)
+    }
+
     fn create_filter(&self, args: CreateFilterArgs) -> Result<FilterInstance, Error> {
         let config: Option<(_, Config)> = args
             .config
@@ -82,7 +86,7 @@ impl FilterFactory for DropFactory {
 }
 
 /// `pass` filter's configuration.
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, schemars::JsonSchema)]
 pub struct Config;
 
 impl TryFrom<proto::Drop> for Config {
